backend/requirements.txt
boto3==1.34.162
botocore==1.34.162
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
jinja2==3.1.4
markdown==3.6
weasyprint==62.3
pydantic==2.8.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
tqdm==4.66.5
openai
googlesearch-python
readability-lxml
opensearch-py